launcher extracted to separate module
