+++ launcher/build.gradle.kts
@@ -33,7 +33,7 @@ group = "io.knotx"
 apply(from = "../gradle/common.deps.gradle.kts")
 apply(from = "../gradle/codegen.deps.gradle.kts")
 dependencies {
-  api(project(":fragment-api"))
+  api(project(":knotx-fragment-api"))
   api(group = "com.google.guava", name = "guava")
   api(group = "commons-io", name = "commons-io")
   api(group = "org.apache.commons", name = "commons-lang3")
@@ -49,6 +49,11 @@ apply(from = "../gradle/common.gradle.kts")
 sourceSets.named("main") {
   java.srcDir("src/main/generated")
 }
+sourceSets.create("junit-test") {
+  compileClasspath = sourceSets.named("main").get().output
+  compileClasspath += sourceSets.named("test").get().runtimeClasspath
+}
+
 
 // -----------------------------------------------------------------------------
 // Tasks
@@ -56,6 +61,19 @@ sourceSets.named("main") {
 
 
 tasks {
+  //FIXME there is race condition with copying Version to generated and compiling project
+  register<Copy>("templatesProcessing") {
+    val tokens = mapOf("project.version" to project.version, "build.timestamp" to "${Utils.timestamp()}")
+    inputs.properties(tokens)
+
+    from("src/main/java-templates") {
+      include("*.java")
+      filter<ReplaceTokens>("tokens" to tokens)
+    }
+    into("src/main/generated/io/knotx/launcher")
+  }
+  getByName<JavaCompile>("compileJava").dependsOn("templatesProcessing")
+
   named<RatTask>("rat") {
     excludes.addAll("**/*.json", "**/*.MD", "**/*.templ", "**/*.adoc", "**/build/*", "**/out/*", "**/generated/*", "/src/test/resources/*", "*.iml")
   }
@@ -82,6 +100,11 @@ tasks.register<Jar>("javadocJar") {
   classifier = "javadoc"
 }
 
+tasks.register<Jar>("testJar") {
+  from(sourceSets.named("junit-test").get().output)
+  classifier = "tests"
+}
+
 publishing {
   publications {
     create<MavenPublication>("mavenJava") {
@@ -89,6 +112,7 @@ publishing {
       from(components["java"])
       artifact(tasks["sourcesJar"])
       artifact(tasks["javadocJar"])
+      artifact(tasks["testJar"])
       pom {
         name.set("Knot.x Launcher")
         description.set("Knot.x Launcher - deploys all Knot.x modules in Vert.x instance.")
