+++ launcher/src/junit-test/java/io/knotx/launcher/rule/KnotxConfiguration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.knotx.junit.rule;
+package io.knotx.launcher.rule;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
