+++ launcher/src/junit-test/java/io/knotx/launcher/rule/TestVertxDeployer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.knotx.junit.rule;
+package io.knotx.launcher.rule;
 
 import com.google.common.collect.Lists;
 import io.knotx.launcher.KnotxStarterVerticle;
