+++ launcher/src/main/java-templates/Version.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.knotx;
+package io.knotx.launcher;
 
 public final class Version {
