+++ launcher/src/main/java/io/knotx/launcher/BadKnotxConfigurationException.java
