+++ launcher/src/main/java/io/knotx/launcher/KnotxCommand.java
