+++ launcher/src/main/java/io/knotx/launcher/KnotxCommandFactory.java
