+++ launcher/src/main/java/io/knotx/launcher/ModuleDescriptor.java
