+++ launcher/src/main/java/io/knotx/launcher/config/ConfProcessor.java
