+++ server-http/api/build.gradle.kts
@@ -33,7 +33,7 @@ group = "io.knotx"
 apply(from = "../../gradle/common.deps.gradle.kts")
 apply(from = "../../gradle/codegen.deps.gradle.kts")
 dependencies {
-  api(project(":fragment-api"))
+  api(project(":knotx-fragment-api"))
   api(group = "com.google.guava", name = "guava")
   api(group = "commons-io", name = "commons-io")
   api(group = "org.apache.commons", name = "commons-lang3")
