+++ settings.gradle.kts
@@ -20,8 +20,7 @@ include("knotx-server-http-api")
 include("knotx-core")
 include("knotx-launcher")
 include("it-test")
-include("documentation")
 
-project("knotx-fragment-api").projectDir = file("fragment-api")
-project("knotx-server-http-api").projectDir = file("server-http/api")
-project("knotx-launcher").projectDir = file("launcher")
+project(":knotx-fragment-api").projectDir = file("fragment-api")
+project(":knotx-server-http-api").projectDir = file("server-http/api")
+project(":knotx-launcher").projectDir = file("launcher")
